Experimenting some different things with menus.
Need to consume MenuItem because of the shortcut and selected
references. Probably need to revert the other options structs to do this
for consistency.

Like the idea of having a no options default func with most commonly
used options, and separate methods or a struct for uncomming options.
It's hard to find a balance though.

Wonder if it would be possible to automatically build via Drop trait.